+++ src/main/java/edu/wpi/rail/jrosbridge/core/Ros.java
@@ -19,6 +19,8 @@ import javax.websocket.OnOpen;
 import javax.websocket.Session;
 
 import edu.wpi.rail.jrosbridge.JRosbridge;
+import edu.wpi.rail.jrosbridge.core.callback.ServiceCallback;
+import edu.wpi.rail.jrosbridge.core.callback.TopicCallback;
 
 /**
  * The Ros object is the main connection point to the rosbridge server. This