+++ src/main/java/edu/wpi/rail/jrosbridge/core/callback/ParamCallback.java
@@ -1,8 +1,8 @@
-package edu.wpi.rail.jrosbridge.core;
+package edu.wpi.rail.jrosbridge.core.callback;
 
 /**
  * The ParamCallback interface defines a single method which will be called when
- * an incoming response is received a rosapi call.
+ * an incoming response is received from rosapi get parameter call.
  * 
  * @author Russell Toris - rctoris@wpi.edu
  * @version Feb. 18, 2014
@@ -12,8 +12,8 @@ public interface ParamCallback {
 	/**
 	 * Handle the response from the rosapi get parameter call.
 	 * 
-	 * @param param
+	 * @param value
 	 *            The value of the returned parameter.
 	 */
-	public void handleResponse(String param);
+	public void handleParameter(String value);
 }