+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 DEVEL - **0.0.3**
  * Support for parameters added [(rctoris)](https://github.com/rctoris/)
- * Math package added [(rctoris)](https://github.com/rctoris/)
+ * `edu.wpi.rail.jrosbridge.math package` added [(rctoris)](https://github.com/rctoris/)
+ * Callbacks added to their own `edu.wpi.rail.jrosbridge.core.callback` package [(rctoris)](https://github.com/rctoris/)
 
 2013-02-16 - **0.0.2**
  * Support for calling services added [(rctoris)](https://github.com/rctoris/)